+++ modeling/configs/maskdino_R50_bs16_50ep_4s_dowsample1_1024.yaml
@@ -25,7 +25,7 @@ MODEL:
     MASK_DIM: 256
     NORM: "GN"
     # pixel decoder
-    PIXEL_DECODER_NAME: "MaskDINOEncoder"
+    PIXEL_DECODER_NAME: "DINOEncoder"
     DIM_FEEDFORWARD: 1024
     NUM_FEATURE_LEVELS: 4
     TOTAL_NUM_FEATURE_LEVELS: 5
@@ -35,7 +35,7 @@ MODEL:
     TRANSFORMER_ENC_LAYERS: 6
     FEATURE_ORDER: "low2high"
   MaskDINO:
-    TRANSFORMER_DECODER_NAME: "MaskDINODecoder"
+    TRANSFORMER_DECODER_NAME: "DINODecoder"
     DEEP_SUPERVISION: True
     NO_OBJECT_WEIGHT: 0.1
     CLASS_WEIGHT: 4.0
