+++ modeling/decoder/__init__.py
@@ -1,3 +1,3 @@
 # Copyright (c) IDEA, Inc. and its affiliates.
-from .maskdino_decoder import MaskDINODecoder
+from .dino_decoder import DINODecoder
