+++ modeling/decoder/dino_decoder.py
@@ -3,269 +3,195 @@
 # Copyright (c) 2022 IDEA. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-# Modified from DINO https://github.com/IDEA-Research/DINO by Feng Li and Hao Zhang.
-# ------------------------------------------------------------------------
-
-from typing import Optional, List, Union
+# Modified from Mask2Former https://github.com/facebookresearch/Mask2Former by Feng Li and Hao Zhang.
+import logging
+import fvcore.nn.weight_init as weight_init
 import torch
-from torch import nn, Tensor
-from torch.cuda.amp import autocast
+from torch import nn
 
-from modeling.utils.utils import MLP, _get_clones, _get_activation_fn, gen_sineembed_for_position, inverse_sigmoid
-from modeling.encoder.ops.modules import MSDeformAttn
+from detectron2.config import configurable
+from detectron2.layers import Conv2d
+
+from .decoder_layer import TransformerDecoder, DeformableTransformerDecoderLayer
+from modeling.utils.utils import MLP, inverse_sigmoid
 
 
-class TransformerDecoder(nn.Module):
-    def __init__(self, decoder_layer, num_layers, norm=None,
-                 return_intermediate=False,
-                 d_model=256, query_dim=4,
-                 modulate_hw_attn=True,
-                 num_feature_levels=1,
-                 deformable_decoder=True,
-                 decoder_query_perturber=None,
-                 dec_layer_number=None,  # number of queries each layer in decoder
-                 rm_dec_query_scale=True,
-                 dec_layer_share=False,
-                 dec_layer_dropout_prob=None,
-                 ):
-        super().__init__()
-        if num_layers > 0:
-            self.layers = _get_clones(decoder_layer, num_layers, layer_share=dec_layer_share)
-        else:
-            self.layers = []
-        self.num_layers = num_layers
-        self.norm = norm
-        self.return_intermediate = return_intermediate
-        assert return_intermediate, "support return_intermediate only"
-        self.query_dim = query_dim
-        assert query_dim in [2, 4], "query_dim should be 2/4 but {}".format(query_dim)
-        self.num_feature_levels = num_feature_levels
-
-        self.ref_point_head = MLP(query_dim // 2 * d_model, d_model, d_model, 2)
-        if not deformable_decoder:
-            self.query_pos_sine_scale = MLP(d_model, d_model, d_model, 2)
-        else:
-            self.query_pos_sine_scale = None
-
-        if rm_dec_query_scale:
-            self.query_scale = None
-        else:
-            raise NotImplementedError
-            self.query_scale = MLP(d_model, d_model, d_model, 2)
-        self.bbox_embed = None
-        self.class_embed = None
-
-        self.d_model = d_model
-        self.modulate_hw_attn = modulate_hw_attn
-        self.deformable_decoder = deformable_decoder
-
-        if not deformable_decoder and modulate_hw_attn:
-            self.ref_anchor_head = MLP(d_model, d_model, 2, 2)
-        else:
-            self.ref_anchor_head = None
-
-        self.decoder_query_perturber = decoder_query_perturber
-        self.box_pred_damping = None
-
-        self.dec_layer_number = dec_layer_number
-        if dec_layer_number is not None:
-            assert isinstance(dec_layer_number, list)
-            assert len(dec_layer_number) == num_layers
-            # assert dec_layer_number[0] ==
-
-        self.dec_layer_dropout_prob = dec_layer_dropout_prob
-        if dec_layer_dropout_prob is not None:
-            assert isinstance(dec_layer_dropout_prob, list)
-            assert len(dec_layer_dropout_prob) == num_layers
-            for i in dec_layer_dropout_prob:
-                assert 0.0 <= i <= 1.0
-
-        self._reset_parameters()
-
-    def _reset_parameters(self):
-        for p in self.parameters():
-            if p.dim() > 1:
-                nn.init.xavier_uniform_(p)
-        for m in self.modules():
-            if isinstance(m, MSDeformAttn):
-                m._reset_parameters()
-
-    def forward(self, tgt, memory,
-                tgt_mask: Optional[Tensor] = None,
-                memory_mask: Optional[Tensor] = None,
-                tgt_key_padding_mask: Optional[Tensor] = None,
-                memory_key_padding_mask: Optional[Tensor] = None,
-                pos: Optional[Tensor] = None,
-                refpoints_unsigmoid: Optional[Tensor] = None,  # num_queries, bs, 2
-                # for memory
-                level_start_index: Optional[Tensor] = None,  # num_levels
-                spatial_shapes: Optional[Tensor] = None,  # bs, num_levels, 2
-                valid_ratios: Optional[Tensor] = None,
-
-                ):
+class DINODecoder(nn.Module):
+    @configurable
+    def __init__(
+            self,
+            in_channels,
+            num_classes: int,
+            hidden_dim: int,
+            dim_feedforward: int,
+            dec_layers: int,
+            enforce_input_project: bool,
+            dn: str,
+            noise_scale: float,
+            dn_num: int,
+            total_num_feature_levels: int = 4,
+            dropout: float = 0.0,
+            activation: str = 'relu',
+            nhead: int = 8,
+            dec_n_points: int = 4,
+            query_dim: int = 4,
+            dec_layer_share: bool = False,
+    ):
         """
-        Input:
-            - tgt: nq, bs, d_model
-            - memory: hw, bs, d_model
-            - pos: hw, bs, d_model
-            - refpoints_unsigmoid: nq, bs, 2/4
-            - valid_ratios/spatial_shapes: bs, nlevel, 2
+        NOTE: this interface is experimental.
+        Args:
+            in_channels: channels of the input features
+            mask_classification: whether to add mask classifier or not
+            num_classes: number of classes
+            hidden_dim: Transformer feature dimension
+            nheads: number of heads
+            dim_feedforward: feature dimension in feedforward network
+            enc_layers: number of Transformer encoder layers
+            dec_layers: number of Transformer decoder layers
+            pre_norm: whether to use pre-LayerNorm or not
+            enforce_input_project: add input project 1x1 conv even if input
+                channels and hidden dim is identical
+            d_model: transformer dimension
+            dropout: dropout rate
+            activation: activation function
+            nhead: num heads in multi-head attention
+            dec_n_points: number of sampling points in decoder
+            query_dim: 4 -> (x, y, w, h)
+            dec_layer_share: whether to share each decoder layer
         """
-        output = tgt
-        device = tgt.device
-
-        intermediate = []
-        reference_points = refpoints_unsigmoid.sigmoid().to(device)
-        ref_points = [reference_points]
-
-        for layer_id, layer in enumerate(self.layers):
-            # preprocess ref points
-            if self.training and self.decoder_query_perturber is not None and layer_id != 0:
-                reference_points = self.decoder_query_perturber(reference_points)
-
-            reference_points_input = reference_points[:, :, None] \
-                                         * torch.cat([valid_ratios, valid_ratios], -1)[None, :]  # nq, bs, nlevel, 4
-            query_sine_embed = gen_sineembed_for_position(reference_points_input[:, :, 0, :]) # nq, bs, 256*2
-
-            raw_query_pos = self.ref_point_head(query_sine_embed)  # nq, bs, 256
-            pos_scale = self.query_scale(output) if self.query_scale is not None else 1
-            query_pos = pos_scale * raw_query_pos
-
-            output = layer(
-                tgt=output,
-                tgt_query_pos=query_pos,
-                tgt_query_sine_embed=query_sine_embed,
-                tgt_key_padding_mask=tgt_key_padding_mask,
-                tgt_reference_points=reference_points_input,
-
-                memory=memory,
-                memory_key_padding_mask=memory_key_padding_mask,
-                memory_level_start_index=level_start_index,
-                memory_spatial_shapes=spatial_shapes,
-                memory_pos=pos,
-
-                self_attn_mask=tgt_mask,
-                cross_attn_mask=memory_mask
-            )
-
-            # iter update
-            if self.bbox_embed is not None:
-                reference_before_sigmoid = inverse_sigmoid(reference_points)
-                delta_unsig = self.bbox_embed[layer_id](output).to(device)
-                outputs_unsig = delta_unsig + reference_before_sigmoid
-                new_reference_points = outputs_unsig.sigmoid()
-
-                reference_points = new_reference_points.detach()
-                # if layer_id != self.num_layers - 1:
-                ref_points.append(new_reference_points)
-
-            intermediate.append(self.norm(output))
-
-        return [
-            [itm_out.transpose(0, 1) for itm_out in intermediate],
-            [itm_refpoint.transpose(0, 1) for itm_refpoint in ref_points]
-        ]
-
-
-class DeformableTransformerDecoderLayer(nn.Module):
-
-    def __init__(self, d_model=256, d_ffn=1024,
-                 dropout=0.1, activation="relu",
-                 n_levels=4, n_heads=8, n_points=4,
-                 use_deformable_box_attn=False,
-                 key_aware_type=None,
-                 ):
         super().__init__()
 
-        # cross attention
-        if use_deformable_box_attn:
-            raise NotImplementedError
-        else:
-            self.cross_attn = MSDeformAttn(d_model, n_levels, n_heads, n_points)
-        self.dropout1 = nn.Dropout(dropout)
-        self.norm1 = nn.LayerNorm(d_model)
+        self.num_feature_levels = total_num_feature_levels
+        # define Transformer decoder here
+        self.dn = dn
+        self.noise_scale = noise_scale
+        self.dn_num = dn_num
+        self.num_layers = dec_layers
+        self.total_num_feature_levels = total_num_feature_levels
 
-        # self attention
-        self.self_attn = nn.MultiheadAttention(d_model, n_heads, dropout=dropout)
-        self.dropout2 = nn.Dropout(dropout)
-        self.norm2 = nn.LayerNorm(d_model)
-
-        # ffn
-        self.linear1 = nn.Linear(d_model, d_ffn)
-        self.activation = _get_activation_fn(activation)
-        self.dropout3 = nn.Dropout(dropout)
-        self.linear2 = nn.Linear(d_ffn, d_model)
-        self.dropout4 = nn.Dropout(dropout)
-        self.norm3 = nn.LayerNorm(d_model)
-
-        self.key_aware_type = key_aware_type
-        self.key_aware_proj = None
-
-    def rm_self_attn_modules(self):
-        self.self_attn = None
-        self.dropout2 = None
-        self.norm2 = None
-
-    @staticmethod
-    def with_pos_embed(tensor, pos):
-        return tensor if pos is None else tensor + pos
-
-    def forward_ffn(self, tgt):
-        tgt2 = self.linear2(self.dropout3(self.activation(self.linear1(tgt))))
-        tgt = tgt + self.dropout4(tgt2)
-        tgt = self.norm3(tgt)
-        return tgt
-
-    @autocast(enabled=False)
-    def forward(self,
-                # for tgt
-                tgt: Optional[Tensor],  # nq, bs, d_model
-                tgt_query_pos: Optional[Tensor] = None,  # pos for query. MLP(Sine(pos))
-                tgt_query_sine_embed: Optional[Tensor] = None,  # pos for query. Sine(pos)
-                tgt_key_padding_mask: Optional[Tensor] = None,
-                tgt_reference_points: Optional[Tensor] = None,  # nq, bs, 4
-
-                # for memory
-                memory: Optional[Tensor] = None,  # hw, bs, d_model
-                memory_key_padding_mask: Optional[Tensor] = None,
-                memory_level_start_index: Optional[Tensor] = None,  # num_levels
-                memory_spatial_shapes: Optional[Tensor] = None,  # bs, num_levels, 2
-                memory_pos: Optional[Tensor] = None,  # pos for memory
-
-                # sa
-                self_attn_mask: Optional[Tensor] = None,  # mask used for self-attention
-                cross_attn_mask: Optional[Tensor] = None,  # mask used for cross-attention
-                ):
-        """
-        Input:
-            - tgt/tgt_query_pos: nq, bs, d_model
-            -
-        """
-        # self attention
-        if self.self_attn is not None:
-            q = k = self.with_pos_embed(tgt, tgt_query_pos)
-            tgt2 = self.self_attn(q, k, tgt, attn_mask=self_attn_mask)[0]
-            tgt = tgt + self.dropout2(tgt2)
-            tgt = self.norm2(tgt)
-
-        # cross attention
-        if self.key_aware_type is not None:
-            if self.key_aware_type == 'mean':
-                tgt = tgt + memory.mean(0, keepdim=True)
-            elif self.key_aware_type == 'proj_mean':
-                tgt = tgt + self.key_aware_proj(memory).mean(0, keepdim=True)
+        self.input_proj = nn.ModuleList()
+        for _ in range(self.num_feature_levels):
+            if in_channels != hidden_dim or enforce_input_project:
+                self.input_proj.append(Conv2d(in_channels, hidden_dim, kernel_size=1))
+                weight_init.c2_xavier_fill(self.input_proj[-1])
             else:
-                raise NotImplementedError("Unknown key_aware_type: {}".format(self.key_aware_type))
-        tgt2 = self.cross_attn(self.with_pos_embed(tgt, tgt_query_pos).transpose(0, 1),
-                               tgt_reference_points.transpose(0, 1).contiguous(),
-                               memory.transpose(0, 1), memory_spatial_shapes, memory_level_start_index,
-                               memory_key_padding_mask).transpose(0, 1)
-        tgt = tgt + self.dropout1(tgt2)
-        tgt = self.norm1(tgt)
+                self.input_proj.append(nn.Sequential())
+        self.num_classes = num_classes
 
-        # ffn
-        tgt = self.forward_ffn(tgt)
+        # init decoder
+        self.decoder_norm = decoder_norm = nn.LayerNorm(hidden_dim)
+        decoder_layer = DeformableTransformerDecoderLayer(hidden_dim, dim_feedforward,
+                                                          dropout, activation,
+                                                          self.num_feature_levels, nhead, dec_n_points)
+        self.decoder = TransformerDecoder(decoder_layer, self.num_layers, decoder_norm,
+                                          d_model=hidden_dim, query_dim=query_dim,
+                                          dec_layer_share=dec_layer_share,
+                                          )
 
-        return tgt
+        self.hidden_dim = hidden_dim
+        _bbox_embed = MLP(hidden_dim, hidden_dim, 4, 3)
+        nn.init.constant_(_bbox_embed.layers[-1].weight.data, 0)
+        nn.init.constant_(_bbox_embed.layers[-1].bias.data, 0)
+        box_embed_layerlist = [_bbox_embed for i in range(self.num_layers)]  # share box prediction each layer
+        self.bbox_embed = nn.ModuleList(box_embed_layerlist)
+        self.decoder.bbox_embed = self.bbox_embed
 
+    @classmethod
+    def from_config(cls, cfg):
+        ret = {}
+        ret["in_channels"] = cfg.MODEL.ENCODER.HIDDEN_DIM
+        ret["num_classes"] = cfg.MODEL.DECODER.NUM_CLASSES
+        ret["hidden_dim"] = cfg.MODEL.DECODER.HIDDEN_DIM
+        # Transformer parameters:
+        ret["dim_feedforward"] = cfg.MODEL.DECODER.DIM_FEEDFORWARD
+        ret["dec_layers"] = cfg.MODEL.DECODER.NUM_LAYERS
+        ret["enforce_input_project"] = cfg.MODEL.DECODER.ENFORCE_INPUT_PROJ
+        ret["dn"]= cfg.MODEL.DECODER.DN
+        ret["noise_scale"] = cfg.MODEL.DECODER.DN_NOISE_SCALE
+        ret["dn_num"] = cfg.MODEL.DECODER.DN_NUM
+        ret["total_num_feature_levels"] = cfg.MODEL.DECODER.TOTAL_NUM_FEATURE_LEVELS
+        return ret
 
+    def pred_box(self, reference, hs, ref0=None):
+        """
+        :param reference: reference box coordinates from each decoder layer
+        :param hs: content
+        :param ref0: whether there are prediction from the first layer
+        """
+        device = reference[0].device
+
+        if ref0 is None:
+            outputs_coord_list = []
+        else:
+            outputs_coord_list = [ref0.to(device)]
+        for dec_lid, (layer_ref_sig, layer_bbox_embed, layer_hs) in enumerate(zip(reference[:-1], self.bbox_embed, hs)):
+            layer_delta_unsig = layer_bbox_embed(layer_hs).to(device)
+            layer_outputs_unsig = layer_delta_unsig + inverse_sigmoid(layer_ref_sig).to(device)
+            layer_outputs_unsig = layer_outputs_unsig.sigmoid()
+            outputs_coord_list.append(layer_outputs_unsig)
+        outputs_coord_list = torch.stack(outputs_coord_list)
+        return outputs_coord_list
+
+    def forward(self, ms_features, target, refpoint_unsigmoid, masks):
+        """
+        :param ms_features: input, a list of multi-scale feature
+            encoder에서 출력하는 multi_scale_features (5 levels) [[1, 256, 200, 304],..., [1, 256, 13, 19]]
+        :param target: query features,[bs, nq, d_model]  It could be from encoder or learnable embedding
+        :param refpoint_unsigmoid:  unsigmoided points or boxes [bs, nq, 2 or 4]
+        :param masks: mask in the original image
+        """
+        assert len(ms_features) == self.num_feature_levels
+        size_list = []
+        # disable mask, it does not affect performance
+        enable_mask = 0
+        if masks is not None:
+            for src in ms_features:
+                if src.size(2) % 32 or src.size(3) % 32:
+                    enable_mask = 1
+        if enable_mask == 0:
+            masks = [torch.zeros((src.size(0), src.size(2), src.size(3)), device=src.device, dtype=torch.bool) for src in ms_features]
+
+        # feature map flatten 해서 하나로 합치고, level_start_index 만들기
+        src_flatten = []
+        mask_flatten = []
+        spatial_shapes = []
+        for i in range(self.num_feature_levels):
+            idx=self.num_feature_levels-1-i
+            size_list.append(ms_features[i].shape[-2:])
+            spatial_shapes.append(ms_features[idx].shape[-2:])
+            src_flatten.append(self.input_proj[idx](ms_features[idx]).flatten(2).transpose(1, 2))
+            mask_flatten.append(masks[i].flatten(1))
+
+        # src_flatten: encoder의 multi-scale feature를 linear projection 해서 한 줄로 합친것
+        src_flatten = torch.cat(src_flatten, 1)  # bs, \sum{hxw}, c
+        mask_flatten = torch.cat(mask_flatten, 1)  # bs, \sum{hxw}
+        spatial_shapes = torch.as_tensor(spatial_shapes, dtype=torch.long, device=src_flatten.device)
+        level_start_index = torch.cat((spatial_shapes.new_zeros((1,)), spatial_shapes.prod(1).cumsum(0)[:-1]))
+
+        # 원래 이 사이에 target과 refpoints_unsigmoid를 만들기 위한 많은 코드가 있었음
+
+        predictions_class = []
+        hs, references = self.decoder(
+            tgt=target.transpose(0, 1),  # query vectors
+            memory=src_flatten.transpose(0, 1),  # key, value from encoder features
+            refpoints_unsigmoid=refpoint_unsigmoid.transpose(0, 1),  # query 좌표를 unsigmoid 한 것 (cx,cy,w,h)
+            memory_key_padding_mask=mask_flatten,
+            level_start_index=level_start_index,  # memory 의 레벨별 feature의 시작 index
+            spatial_shapes=spatial_shapes,  # memory에 들어있는 feature shapes
+        )                      # MultiheadAttention.forward() 참조
+        # hs: output features of decoder layers
+        # references: output points of decoder layers (refined points from query)
+
+        for i, output in enumerate(hs):
+            # decoder_norm: LayerNorm
+            outputs_class = self.decoder_norm(output.transpose(0, 1))
+            outputs_class = outputs_class.transpose(0, 1)
+            predictions_class.append(outputs_class)
+        out_boxes = self.pred_box(references, hs)
+
+        # 최종 출력은 마지막 레이어의 출력만 내보낸다.
+        out = {'pred_logit': predictions_class[-1],
+               'pred_box': out_boxes[-1],
+               'pred_feat': hs[-1]
+              }
+        return out
