+++ modeling/encoder/__init__.py
@@ -1,1 +1,2 @@
 # Copyright (c) IDEA, Inc. and its affiliates.
+from .dino_encoder import DINOEncoder
